+++ asusctl/examples/aura-zoned-breathe.rs
@@ -1,6 +1,6 @@
 //! Using a combination of key-colour array plus a key layout to generate outputs.
 
-use rog_aura::{layouts::KeyLayout, ActionData, Colour, LedType, PerZone, Sequences, Speed};
+use rog_aura::{layouts::KeyLayout, Colour, LedType, PerZone, Sequences, Speed, Effect, Breathe};
 use rog_dbus::RogDbusClientBlocking;
 
 fn main() -> Result<(), Box<dyn std::error::Error>> {
@@ -10,28 +10,31 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
 
     let mut seq = Sequences::new();
 
-    let zone = ActionData::new_breathe(
+    let zone = Effect::Breathe(
+        Breathe::new(
         LedType::Zone(PerZone::KeyboardLeft),
         Colour(166, 127, 166),
         Colour(127, 155, 20),
         Speed::High,
-    );
+    ));
     seq.push(zone);
 
-    let zone = ActionData::new_breathe(
+    let zone = Effect::Breathe(
+        Breathe::new(
         LedType::Zone(PerZone::KeyboardCenterLeft),
         Colour(16, 127, 255),
         Colour(127, 15, 20),
         Speed::Low,
-    );
+    ));
     seq.push(zone);
 
-    let zone = ActionData::new_breathe(
+    let zone = Effect::Breathe(
+        Breathe::new(
         LedType::Zone(PerZone::LightbarRightCorner),
         Colour(0, 255, 255),
         Colour(255, 0, 255),
         Speed::Med,
-    );
+    ));
     seq.push(zone);
 
     loop {